+++ packages/gatsby-plugin-sharp/src/gatsby-node.js
@@ -21,6 +21,19 @@ try {
   coreSupportsOnPluginInit = false
 }
 
+function removeCachedValue(cache, key) {
+  if (cache?.del) {
+    // if cache expose ".del" method directly on public interface
+    return cache.del(key)
+  } else if (cache?.cache?.del) {
+    // legacy - using internal cache instance and calling ".del" on it directly
+    return cache.cache.del(key)
+  }
+  return Promise.reject(
+    new Error(`Cache instance doesn't expose ".del" function`)
+  )
+}
+
 exports.onCreateDevServer = async ({ app, cache, reporter }) => {
   if (!_lazyJobsEnabled()) {
     return
@@ -53,14 +66,14 @@ exports.onCreateDevServer = async ({ app, cache, reporter }) => {
     } = splitOperationsByRequestedFile(cacheResult, pathOnDisk)
 
     await _unstable_createJob(matchingJob, { reporter })
-    await cache.cache.del(decodedURI)
+    await removeCachedValue(cache, decodedURI)
 
     if (jobWithRemainingOperations.args.operations.length > 0) {
       // There are still some operations pending for this job - replace the cached job
-      await cache.cache.set(jobContentDigest, jobWithRemainingOperations)
+      await cache.set(jobContentDigest, jobWithRemainingOperations)
     } else {
       // No operations left to process - purge the cache
-      await cache.cache.del(jobContentDigest)
+      await removeCachedValue(cache, jobContentDigest)
     }
 
     return res.sendFile(pathOnDisk)