+++ packages/gatsby/src/utils/cache-lmdb.ts
@@ -1,4 +1,4 @@
-import { open, RootDatabase, Database } from "lmdb-store"
+import { open, RootDatabase, Database, DatabaseOptions } from "lmdb-store"
 import fs from "fs-extra"
 import path from "path"
 
@@ -16,13 +16,26 @@ const cacheDbFile =
 export default class GatsbyCacheLmdb {
   private static store
   private db: Database | undefined
+  private encoding: DatabaseOptions["encoding"]
   public readonly name: string
   // Needed for plugins that want to write data to the cache directory
   public readonly directory: string
+  // TODO: remove `.cache` in v4. This is compat mode - cache-manager cache implementation
+  // expose internal cache that gives access to `.del` function that wasn't available in public
+  // cache interface (gatsby-plugin-sharp use it to clear no longer needed data)
+  public readonly cache: GatsbyCacheLmdb
 
-  constructor({ name = `db` }: { name: string }) {
+  constructor({
+    name = `db`,
+    encoding = `json`,
+  }: {
+    name: string
+    encoding?: DatabaseOptions["encoding"]
+  }) {
     this.name = name
+    this.encoding = encoding
     this.directory = path.join(process.cwd(), `.cache/caches/${name}`)
+    this.cache = this
   }
 
   init(): GatsbyCacheLmdb {
@@ -46,6 +59,7 @@ export default class GatsbyCacheLmdb {
     if (!this.db) {
       this.db = GatsbyCacheLmdb.getStore().openDB({
         name: this.name,
+        encoding: this.encoding,
       })
     }
     return this.db
@@ -58,5 +72,9 @@ export default class GatsbyCacheLmdb {
   async set<T>(key: string, value: T): Promise<T | undefined> {
     await this.getDb().put(key, value)
     return value
+  }
+
+  async del(key: string): Promise<void> {
+    return (this.getDb().remove(key) as unknown) as Promise<void>
   }
 }