+++ packages/gatsby/src/utils/worker/child/index.ts
@@ -7,5 +7,5 @@ initReporterMessagingInWorker()
 // Note: this doesn't check for conflicts between module exports
 export { renderHTMLProd, renderHTMLDev } from "./render-html"
 export { setInferenceMetadata, buildSchema } from "./schema"
-export { setComponents, runQueries, saveQueries } from "./queries"
+export { setComponents, runQueries, saveQueriesDependencies } from "./queries"
 export { loadConfigAndPlugins } from "./load-config-and-plugins"